"Extract SectionDesign to a separate file for better modularity in Shell design. Remove duplicate definition from ShellDesign.cs."
